requirements.txt
# ---------------------------
# ✅ CPU-compatible PyTorch (no CUDA)
# ---------------------------
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

# ---------------------------
# ✅ EasyOCR and dependencies
# ---------------------------
easyocr==1.7.1
opencv-python-headless
numpy==1.24.4
Pillow
scipy
scikit-image
pyclipper
shapely
python-bidi
arabic-reshaper

# ---------------------------
# ✅ Streamlit
# ---------------------------
streamlit
